f32 -> i16

+++ src/engine.rs
@@ -12,10 +12,10 @@ use web_sys::{CanvasRenderingContext2d, HtmlImageElement};
 use crate::browser::{self, LoopClosure};
 
 pub struct Rect {
-    pub x: f32,
-    pub y: f32,
-    pub width: f32,
-    pub height: f32,
+    pub x: i16,
+    pub y: i16,
+    pub width: i16,
+    pub height: i16,
 }
 
 impl Rect {
@@ -235,8 +235,8 @@ impl Image {
         let bounding_box = Rect {
             x: position.x.into(),
             y: position.y.into(),
-            width: element.width() as f32,
-            height: element.height() as f32,
+            width: element.width() as i16,
+            height: element.height() as i16,
         };
 
         Self {
@@ -259,7 +259,7 @@ impl Image {
     }
 
     pub fn set_x(&mut self, x: i16) {
-        self.bounding_box.x = x as f32;
+        self.bounding_box.x = x;
         self.position.x = x;
     }
